refactor(server): split span analytics tools module by concern
The tools module carried four separable concerns beside the tool
definitions themselves: response-envelope schemas and size budgeting,
observed-path discovery sampling, zero-result diagnosis, and Phoenix UI
link composition. Each moves to its own module (envelope, discovery,
links) with the tool definitions, parameter schemas, and handlers
remaining in tools. The public entry point
(build_span_analytics_tools) and all behavior are unchanged; the
request accessor is injected into link composition so it stays
independent of the serving stack.
