+++ src/phoenix/server/mcp_span_analytics/__init__.py
@@ -2,8 +2,11 @@
 
 See :mod:`phoenix.server.mcp_span_analytics.tools` for the tool surface,
 :mod:`phoenix.server.mcp_span_analytics.registry` for the field registry,
-and :mod:`phoenix.server.mcp_span_analytics.compiler` for query
-compilation.
+:mod:`phoenix.server.mcp_span_analytics.compiler` for query compilation,
+:mod:`phoenix.server.mcp_span_analytics.discovery` for observed-path
+sampling, :mod:`phoenix.server.mcp_span_analytics.envelope` for response
+envelopes and size budgeting, and
+:mod:`phoenix.server.mcp_span_analytics.links` for UI link composition.
 """
 
 from phoenix.server.mcp_span_analytics.tools import build_span_analytics_tools
